Move CancelContext into ServiceContext
