+++ dubbo-rpc/dubbo-rpc-api/src/main/java/org/apache/dubbo/rpc/CancellationListener.java
@@ -27,6 +27,6 @@ public interface CancellationListener {
      *
      * @param context the newly cancelled context.
      */
-    void cancelled(CancellableContext context);
+    void cancelled(RpcServiceContext context);
 
 }
